Fixing of building errors
